images/alpine: pin base image version
Take the opportunity to use Alpine Edge (which is what these dated tags
track), instead of Alpine Stable (which is what the "latest" tag tracks)
because Alpine Edge gets security fixes (for many important libraries
that programs like Git depend on) on a much faster cadence, in exchange
for less overall stability. For our use case we care more about security
than stability, so this is a fair trade for us.

One benefit of pinning is that we can easily bump these images with a
new date, which makes things more obvious in the bump PRs, which
technically only need to touch some other part of the directory (any
file under images/alpine/* for example) for Prow to notice and run the
image publishing jobs as a postsubmit.