images/git-custom-k8s-auth: pin gcloud sdk version
This way, we can update the version of aws-iam-authenticator without
bumping gke-gcloud-auth-plugin, and vice versa. For example, in
https://github.com/kubernetes/test-infra/pull/31725 we bumped only
aws-iam-authenticator but if we look at the image build history in GCR
we have the following tags:

    v20240125-32d3286cca (built with #31725 just yesterday)
    v20230324-76cde35b3d (built 10 months ago...!)
    ...

and there must have been some gcloud SDK changes in those 10 months.
Being explicit about those changes would have been informative. So,
let's start tracking those changes.

Also add an explicit note purposely staying on the 0.5.x series for
aws-iam-authenticator (there is a 0.6.x series in existence which we are
deliberately avoiding, presumably because 0.6.x has breaking changes
we're not ready for).